Moved CForeign check to rxfs_storeInit/storeMore.
This is rxfs-specific, so move it out of CacheStoreProc.

It must be done before entering the do { } while (moredata) loop proper
now, as rxfs_fetchMore is never called during the first iteration

(moredata is always 0), but the length may need fixing (note that
this was done after what is now the fetchMore op originally). Thus
it appears in rxfs_fetchInit.
To prevent a bogus call during the first loop iteration, also check
for length!=0. Finally, don't bother with fetchMore at all unless
CForeign is set.

In rxfs_fetchInit, we added an additional check for length_hi being
zero.

uploade patchset 8.

Reviewed-on: http://gerrit.openafs.org/126
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>