Windows: Use "Name not found" instead of "File not found"
When a file cannot be opened because the name of the file or
a directory in the path cannot be found, return "Name not found"
instead of "File not found".  "File not found" is the error returned
by the smb redirector to the application when the file cannot be
located in the search path.  "Name not found" indicates that the
entry is not present at the location requested.

The use of "File not found" prevented execution of applications
from the \\AFS name space on Windows 7.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/469
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>